[assets] rename directory: accountorcommunity -> account_or_community
