+++ lib/page/assets/account_or_community/AccountOrCommunityData.dart
