+++ lib/page/assets/account_or_community/accountOrCommunityItemHorizontal.dart
