+++ lib/page/assets/account_or_community/switchAccountOrCommunity.dart
